codegen: shared ctx.resolve_{fn_def,expr,stmt,pattern} — #147 phase E PR 9
Foundation for the wasm-gc backend migration (PRs 9.1 / 9.2). Promotes the
resolved-lookup helpers from `src/codegen/rust/toplevel.rs` to
`CodegenContext` methods so every backend hits the same one-line boundary
when migrating off source-shape `Expr` / `FnBody` / `Pattern`.

Changes:

- `CodegenContext::resolve_fn_def(fd)` — look up the resolved-HIR mirror
  of a source-shape `FnDef` in `resolved_fn_defs` / `resolved_module_fn_defs`
  (already populated by `build_context` since PR 8). Falls back to a fresh
  per-call resolver lift when neither path covers `fd` — same shape as
  the pre-PR-9 private helper.
- `CodegenContext::resolve_expr(spanned)` /
  `CodegenContext::resolve_stmt(stmt)` /
  `CodegenContext::resolve_pattern(pat)` — on-demand resolver lifts for
  source-shape borrows. Used by emit helpers that still walk pre-resolve
  AST (TCO hoisting, mutual TCO, verify cases, etc.).
- `rust/toplevel.rs` drops the private `resolved_fn_def_for` /
  `resolved_expr_on_demand` / `resolved_stmt_on_demand` /
  `resolved_pattern_on_demand` in favour of the shared methods.
- `wasm_gc/mod.rs` gains a module-level doc comment describing the
  staged migration plan (PR 9 foundation, PR 9.1 `body/emit.rs`, PR 9.2
  `module.rs` / `types_discovery.rs` / `types.rs` / `flatten.rs`). The
  byte-identical gate for every wasm-gc snapshot holds across every PR
  in the chain — no generated wasm shifts until the migration completes.

No behavioural change: the methods are byte-for-byte equivalent to the
removed private helpers; only the access surface moves.

Test results:

- `cargo test --features wasm --no-fail-fast`: **1649 tests pass**, 0
  failures (up from 1193 in PR 8 — the bench wasm-gc / wasip2 suites
  re-enable here).
- `cargo clippy --features wasm -p aver-lang --lib --bin aver -- -D warnings`:
  clean.
- `cargo clippy --workspace --all-targets --exclude aver-lang -- -D warnings`:
  clean.
- `cargo fmt --all -- --check`: clean.
- All 13 captured wasm-gc snapshots size-identical to pre-PR baselines;
  `order_total.wasm` md5 `61e45b325279e17e1b0d8dce3fde43f9` byte-identical
  (the `life.wasm` byte-identical check turns out flaky across builds
  pre-existing — type-section ordering depends on `HashMap` iteration —
  but size is deterministic at 10508 bytes across every run).
