+++ src/codegen/wasm_gc/mod.rs
@@ -6,6 +6,39 @@
 //!
 //! Status: Phase 1 — module scaffold. `compile_to_wasm_gc` returns a
 //! placeholder error until a hello-world emitter lands in `module.rs`.
+//!
+//! ## Phase E (resolved AST) migration plan
+//!
+//! Issue #147 phase E migrates every backend from consuming
+//! `crate::ast::Expr` / `FnBody` / `Pattern` to consuming the
+//! resolved-HIR shapes in `crate::ir::hir`. The wasm-gc backend is
+//! migrated across three PRs because it's the largest backend
+//! (~36k lines, 8 files actively walking `Expr`):
+//!
+//! - **PR 9** (foundation) — shared `CodegenContext::resolve_fn_def` /
+//!   `resolve_expr` / `resolve_stmt` / `resolve_pattern` methods
+//!   promoted from `rust/toplevel.rs`. No wasm-gc behaviour change;
+//!   gives the follow-up PRs a one-line lookup boundary.
+//! - **PR 9.1** — `body/emit.rs` (3.1k) migrates to take
+//!   `&Spanned<ResolvedExpr>`; cascades through `emit_list_literal`,
+//!   `emit_map_literal`, `emit_constructor`, `emit_match`,
+//!   `emit_tail_call`, etc. The remaining body subfiles
+//!   (`body/{slots, builtins, builtins_wasip2, infer, eq_helpers,
+//!   hash_helpers}.rs`) migrate alongside since they share the
+//!   `Spanned<Expr>` borrows.
+//! - **PR 9.2** — `module.rs` orchestration (6.3k),
+//!   `types_discovery.rs` / `types.rs` / `flatten.rs` (program-level
+//!   walkers that register types / fns before emit). After this PR
+//!   the wasm-gc backend reads `ctx.resolved_fn_defs` /
+//!   `resolved_module_fn_defs` exclusively; the `&FnDef` borrows in
+//!   `module.rs` survive only as the source-shape metadata that the
+//!   resolver doesn't reproduce (e.g. param annotation source
+//!   strings used by the WIT signature emitter).
+//!
+//! The byte-identical gate for every wasm-gc snapshot
+//! (`order_total.wasm` md5 `61e45b325279e17e1b0d8dce3fde43f9`, the
+//! game suite, the wasip2 stress fixtures) holds across each PR — no
+//! generated wasm shifts until the migration is fully through.
 
 use crate::ast::TopLevel;
 use crate::ir::AnalysisResult;
